requirements.txt
absl-py==2.0.0
arviz==0.16.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==3.0.0
cons==0.4.6
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
emcee==3.1.4
etuples==0.3.9
exceptiongroup==1.2.0
fastprogress==1.0.3
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
graphviz==0.20.1
grpcio==1.60.0
h5netcdf==1.3.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
jax==0.4.23
jaxlib==0.4.23
Jinja2==3.1.2
kiwisolver==1.4.5
logical-unification==0.4.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
miniKanren==1.0.3
ml-dtypes==0.3.1
mpmath==1.3.0
multipledispatch==1.0.0
networkx==3.2.1
neurodiffeq==0.6.2
numpy==1.26.2
numpyro==0.13.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pymc==5.10.3
pyparsing==3.1.1
pytensor==2.18.4
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
xarray==2023.12.0
xarray-einstats==0.6.0
zipp==3.17.0
